DjangoCourseHW/pyproject.toml
[tool.poetry]
name = "djangocoursehw"
version = "0.1.0"
description = ""
authors = ["ann-bereza <annhavelia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
requests = "^2.28.2"
markdown = "^3.4.3"
psycopg2-binary = "2.9.6"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
pytz = "^2023.3"
geopy = "^2.3.0"
timezonefinder = "^6.2.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
